Rename CreateGameService to ChooseGameService and add JoinGameService functions; Delete JoinGameService.
