+++ app/src/main/java/com/d/tickettoride/command/client/CBeginPlayCommand.kt
@@ -2,7 +2,7 @@ package com.d.tickettoride.command.client
 
 import com.d.tickettoride.model.Game
 import com.d.tickettoride.service.BeginPlayService
-import com.d.tickettoride.service.CreateGameService
+import com.d.tickettoride.service.ChooseGameService
 import com.d.tickettoride.service.ErrorMessageService
 
 class CBeginPlayCommand(private val game: Game?, private val errorMessage:String?) : ICommand {
@@ -11,7 +11,7 @@ class CBeginPlayCommand(private val game: Game?, private val errorMessage:String
         //BeginPlayService.instance.setCurrentGame(game)
         if (errorMessage == null) {
             BeginPlayService.instance.startGame(true)
-            CreateGameService.instance.removeGameFromList(game!!.info)
+            ChooseGameService.instance.removeGameFromList(game!!.info)
         }
         else {
             ErrorMessageService.instance.postErrorMessage(errorMessage)
