+++ app/src/main/java/com/d/tickettoride/command/client/CJoinGameCommand.kt
@@ -1,13 +1,13 @@
 package com.d.tickettoride.command.client
 
+import com.d.tickettoride.service.ChooseGameService
 import com.d.tickettoride.service.ErrorMessageService
-import com.d.tickettoride.service.JoinGameService
 
 class CJoinGameCommand (private val errorMessage:String?) : ICommand {
 
     override fun execute() {
         if (errorMessage == null) {
-            JoinGameService.instance.setJoinedGame(true)
+            ChooseGameService.instance.setJoinedGame(true)
         }
         else {
             ErrorMessageService.instance.postErrorMessage(errorMessage)
