+++ app/src/main/java/com/d/tickettoride/command/client/CRemoveGameCommand.kt
@@ -1,11 +1,11 @@
 package com.d.tickettoride.command.client
 
 import com.d.tickettoride.model.GameInfo
-import com.d.tickettoride.service.CreateGameService
+import com.d.tickettoride.service.ChooseGameService
 
 class CRemoveGameCommand(private val gameInfo: GameInfo) : ICommand {
 
     override fun execute() {
-        CreateGameService.instance.removeGameFromList(gameInfo)
+        ChooseGameService.instance.removeGameFromList(gameInfo)
     }
 }
