+++ app/src/main/java/com/d/tickettoride/service/ChooseGameService.kt
@@ -1,6 +1,8 @@
 package com.d.tickettoride.service
 
 import com.d.tickettoride.command.server.SCreateGameCommand
+import com.d.tickettoride.command.server.SJoinGameCommand
+import com.d.tickettoride.model.Game
 import com.d.tickettoride.model.GameInfo
 import com.d.tickettoride.model.RootModel
 import com.d.tickettoride.servercommunicator.CommandType
@@ -8,10 +10,10 @@ import com.d.tickettoride.servercommunicator.Poller
 import com.d.tickettoride.servercommunicator.ServerProxy
 import com.google.gson.Gson
 
-class CreateGameService(private val proxy: ServerProxy = ServerProxy()) {
+class ChooseGameService(private val proxy: ServerProxy = ServerProxy()) {
 
     companion object {
-        val instance = CreateGameService()
+        val instance = ChooseGameService()
     }
 
     fun createGame(gameName:String, numPlayers:Int, creator:String) {
@@ -30,6 +32,19 @@ class CreateGameService(private val proxy: ServerProxy = ServerProxy()) {
         rootModel.gameToRemoveFromList = gameInfo
         rootModel.gameList.remove(gameInfo)
         rootModel.gameListLength--
+    }
+
+    fun joinGame(gameName:String) {
+        val data = Gson().toJson(SJoinGameCommand(gameName, RootModel.instance.user!!.userName))
+        ServerProxy().command(CommandType.S_JOIN_GAME, data)
+    }
+
+    fun setJoinedGame(joined: Boolean) {
+        RootModel.instance.waitingForGame = joined
+    }
+
+    fun setGameData(game: Game) {
+        RootModel.instance.game = game
     }
 
     fun startPoller() {
